Generamos un custom hook para la galeria con el objetivo de limpiar mas el codigo del componente
