+++ src/components/core/Button/Button.jsx
@@ -1,13 +1,10 @@
-import { useState } from 'react'
 import './Button.css'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 
 function Button ({ text, icon, img, imgAlt, handleClick, disabled, modifire }) {
-  const [isDisabled] = useState(disabled || false)
-
   return (
     <>
-      <button className={`Button Button--${modifire}`} onClick={handleClick} disabled={isDisabled}>
+      <button className={`Button Button--${modifire}`} onClick={handleClick} disabled={disabled}>
         {icon && (<FontAwesomeIcon icon={icon} className={`Button-icon Button-icon--${modifire}`} />)}
         {img && (<img src={img} alt={imgAlt} className={`Button-img Button-img--${modifire}`} />)}
         {text && (<span className={`Button-text Button-text--${modifire}`}>{text}</span>)}
